usm: tests: classification: Move AMQP classification test to USM linux file
