+++ pkg/network/tracer/tracer_usm_linux_test.go
@@ -37,6 +37,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/network/config"
 	netlink "github.com/DataDog/datadog-agent/pkg/network/netlink/testutil"
 	"github.com/DataDog/datadog-agent/pkg/network/protocols"
+	"github.com/DataDog/datadog-agent/pkg/network/protocols/amqp"
 	"github.com/DataDog/datadog-agent/pkg/network/protocols/http"
 	"github.com/DataDog/datadog-agent/pkg/network/protocols/http/testutil"
 	"github.com/DataDog/datadog-agent/pkg/network/protocols/kafka"
@@ -57,6 +58,7 @@ const (
 	postgresPort = "5432"
 	mongoPort    = "27017"
 	redisPort    = "6379"
+	amqpPort     = "5672"
 	kafkaPort    = "9092"
 
 	produceAPIKey = 0
@@ -1594,6 +1596,135 @@ func testRedisProtocolClassification(t *testing.T, tr *Tracer, clientHost, targe
 			},
 			teardown:   redisTeardown,
 			validation: validateProtocolConnection(&protocols.Stack{Application: protocols.Redis}),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			testProtocolClassificationInner(t, tt, tr)
+		})
+	}
+}
+
+func testAMQPProtocolClassification(t *testing.T, tr *Tracer, clientHost, targetHost, serverHost string) {
+	skipFunc := composeSkips(skipIfNotLinux, skipIfUsingNAT)
+	skipFunc(t, testContext{
+		serverAddress: serverHost,
+		serverPort:    amqpPort,
+		targetAddress: targetHost,
+	})
+
+	defaultDialer := &net.Dialer{
+		LocalAddr: &net.TCPAddr{
+			IP: net.ParseIP(clientHost),
+		},
+	}
+
+	amqpTeardown := func(t *testing.T, ctx testContext) {
+		client := ctx.extras["client"].(*amqp.Client)
+		defer client.Terminate()
+
+		require.NoError(t, client.DeleteQueues())
+	}
+
+	// Setting one instance of amqp server for all tests.
+	serverAddress := net.JoinHostPort(serverHost, amqpPort)
+	targetAddress := net.JoinHostPort(targetHost, amqpPort)
+	require.NoError(t, amqp.RunServer(t, serverHost, amqpPort))
+
+	tests := []protocolClassificationAttributes{
+		{
+			name: "connect",
+			context: testContext{
+				serverPort:    amqpPort,
+				serverAddress: serverAddress,
+				targetAddress: targetAddress,
+				extras:        make(map[string]interface{}),
+			},
+			postTracerSetup: func(t *testing.T, ctx testContext) {
+				client, err := amqp.NewClient(amqp.Options{
+					ServerAddress: ctx.serverAddress,
+					Dialer:        defaultDialer,
+				})
+				require.NoError(t, err)
+				ctx.extras["client"] = client
+			},
+			teardown:   amqpTeardown,
+			validation: validateProtocolConnection(&protocols.Stack{Application: protocols.AMQP}),
+		},
+		{
+			name: "declare channel",
+			context: testContext{
+				serverPort:    amqpPort,
+				serverAddress: serverAddress,
+				targetAddress: targetAddress,
+				extras:        make(map[string]interface{}),
+			},
+			preTracerSetup: func(t *testing.T, ctx testContext) {
+				client, err := amqp.NewClient(amqp.Options{
+					ServerAddress: ctx.serverAddress,
+					Dialer:        defaultDialer,
+				})
+				require.NoError(t, err)
+				ctx.extras["client"] = client
+			},
+			postTracerSetup: func(t *testing.T, ctx testContext) {
+				client := ctx.extras["client"].(*amqp.Client)
+				require.NoError(t, client.DeclareQueue("test", client.PublishChannel))
+			},
+			teardown:   amqpTeardown,
+			validation: validateProtocolConnection(&protocols.Stack{}),
+		},
+		{
+			name: "publish",
+			context: testContext{
+				serverPort:    amqpPort,
+				serverAddress: serverAddress,
+				targetAddress: targetAddress,
+				extras:        make(map[string]interface{}),
+			},
+			preTracerSetup: func(t *testing.T, ctx testContext) {
+				client, err := amqp.NewClient(amqp.Options{
+					ServerAddress: ctx.serverAddress,
+					Dialer:        defaultDialer,
+				})
+				require.NoError(t, err)
+				require.NoError(t, client.DeclareQueue("test", client.PublishChannel))
+				ctx.extras["client"] = client
+			},
+			postTracerSetup: func(t *testing.T, ctx testContext) {
+				client := ctx.extras["client"].(*amqp.Client)
+				require.NoError(t, client.Publish("test", "my msg"))
+			},
+			teardown:   amqpTeardown,
+			validation: validateProtocolConnection(&protocols.Stack{Application: protocols.AMQP}),
+		},
+		{
+			name: "consume",
+			context: testContext{
+				serverPort:    amqpPort,
+				serverAddress: serverAddress,
+				targetAddress: targetAddress,
+				extras:        make(map[string]interface{}),
+			},
+			preTracerSetup: func(t *testing.T, ctx testContext) {
+				client, err := amqp.NewClient(amqp.Options{
+					ServerAddress: ctx.serverAddress,
+					Dialer:        defaultDialer,
+				})
+				require.NoError(t, err)
+				require.NoError(t, client.DeclareQueue("test", client.PublishChannel))
+				require.NoError(t, client.DeclareQueue("test", client.ConsumeChannel))
+				require.NoError(t, client.Publish("test", "my msg"))
+				ctx.extras["client"] = client
+			},
+			postTracerSetup: func(t *testing.T, ctx testContext) {
+				client := ctx.extras["client"].(*amqp.Client)
+				res, err := client.Consume("test", 1)
+				require.NoError(t, err)
+				require.Equal(t, []string{"my msg"}, res)
+			},
+			teardown:   amqpTeardown,
+			validation: validateProtocolConnection(&protocols.Stack{Application: protocols.AMQP}),
 		},
 	}
 	for _, tt := range tests {
